requirements.txt
asgiref==3.2.10
astroid==2.4.2
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.6.0
dj-rest-auth==1.1.2
Django==3.1.2
django-allauth==0.43.0
django-cors-headers==3.5.0
django-filter==2.4.0
django-simple-history==2.12.0
django-storages==1.10.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
flake8==3.8.4
google-api-core==1.23.0
google-auth==1.22.1
google-cloud-core==1.4.3
google-cloud-secret-manager==2.0.0
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.32.0
gunicorn==20.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.5.1
libcst==0.3.13
Markdown==3.3.2
mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.0
Pillow==8.0.0
pkg-resources==0.0.0
proto-plus==1.11.0
protobuf==3.13.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
snowballstemmer==2.0.0
sqlparse==0.4.1
toml==0.10.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.11
wrapt==1.12.1
